scripts/requirements.txt
alabaster==0.7.12
setuptools==68.0.0
amqp==5.0.6
aniso8601==9.0.1
apimatic-core==0.2.8
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.6
argh==0.26.2
astroid==2.15.4
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
CacheControl==0.12.14
celery==5.1.1
certifi==2020.11.8
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.5
cryptography==42.0.5
deprecation==2.1.0
docutils==0.19
enum34==1.1.10
et-xmlfile==1.1.0
exceptiongroup==1.1.1
expiringdict==1.2.1
flake8==3.8.4
flake8-polyfill==1.0.2
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
geographiclib==2.0
geopy==2.4.0
googlemaps==4.10.0
h11==0.14.0
html2text==2020.1.16
humanize==3.9.0
idna==2.10
imagesize==1.2.0
iniconfig==2.0.0
isodate==0.6.1
isort==5.6.4
itsdangerous==1.1.0
jenkinsapi==0.3.11
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==3.0.3
jsonpointer==2.4
jsonschema==3.2.0
kombu==5.2.2
lazy-object-proxy==1.4.3
lxml==4.9.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==10.2.0
msgpack==1.0.8
mysql-connector-python==8.0.33
nameparser==1.1.3
numpy==1.21.2
openpyxl==3.1.2
outcome==1.2.0
packaging==20.4
pandas==2.0.1
pathtools==0.1.2
pdf-tools==0.2.0
pendulum==3.0.0
pep257==0.7.0
pep8-naming==0.10.0
Pillow==9.5.0
pkcs7==0.1.2
platformdirs==4.2.0
pluggy==1.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.20.3
py==1.11.0
py4j==0.10.9
pycodestyle==2.6.0
pycparser==2.21
pycryptodome==3.9.9
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
PyMuPDF==1.22.5
pyparsing==2.4.7
PyPDF2==3.0.1
pyrsistent==0.17.3
PySocks==1.7.1
pyspark==3.1.2
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-google-places==1.4.2
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.10.0
scipy==1.7.3
selenium==4.4.1
simple-salesforce==1.12.5
simplejson==3.19.2
six==1.15.0
sniffio==1.3.0
snowballstemmer==2.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
squareup==35.0.0.20240222
stripe==5.4.0
time-machine==2.14.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
ua-parser==0.10.0
unicodecsv==0.14.1
urllib3==1.26.2
uwsgitop==0.11
vine==5.0.0
watchdog==0.10.3
wcwidth==0.2.5
webdriver-manager
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==1.3.7
xmltodict==0.13.0
zeep==4.2.1
pyap==0.3.1
boto3==1.34.145
icalendar==5.0.13
